Viewer: fix incorrect jump in time when double clicking in some views. Regression introduced in 295435ee85812f073

+++ server/viewer/vwMainPlot.cpp
@@ -655,7 +655,7 @@ vwMain::drawPlot(int curPlotWindowIdx)
             else if(elem.nameIdx==elem.hlNameIdx) newTimeRangeNs = (s64)(vwConst::DCLICK_RANGE_FACTOR*pcp.value); // For scopes, the value is the duration
             else newTimeRangeNs = vwConst::DCLICK_RANGE_FACTOR*cmGetParentDurationNs(_record, pcp.evt.threadId, nestingLevel, pcp.lIdx); // For "flat" items, the duration is the one of the parent
             if(newTimeRangeNs>0.) {
-                s64 newStartTimeNs = bsMax(pw.startTimeNs+(s64)((pcp.timeNs-pw.startTimeNs)/pw.timeRangeNs*(pw.timeRangeNs-newTimeRangeNs)), 0LL);
+                s64 newStartTimeNs = bsMax(pw.startTimeNs+(s64)((double)(pcp.timeNs-pw.startTimeNs)/(double)pw.timeRangeNs*(double)(pw.timeRangeNs-newTimeRangeNs)), 0LL);
                 pw.setView(newStartTimeNs, newTimeRangeNs);
                 changedNavigation = true;
             }
